TST: move concurrency tests to dedicated module, add concurrency test for inifix-validate
